etmem: add sysmem_threshold and swap_threshold parameters to etmem
As a memory expansion tool, etmem performs memory swap operation
for the target process by default after it is enabled. However,
in some specific scenarios, in order to obtain the ultimate
performance of the business, it is necessary to consider
the timing of memory swapping by etmem.

Add the sysmem_threshold parameter to control the start and
stop of memory swap out when the system available memory
meets the requirements.

Add the swap_threshold parameter to limit the absolute
value of the memory size reserved by the process in DRAM.
When the target process memory is lower than this value, do not
swap out any memory.

Signed-off-by: liubo <liubo254@huawei.com>