etmem: add swap flag to support specified page swap out
The existing memory expansion tool etmem swaps out
all pages that can be swapped out for the process by
default, unless the page is marked with lock flag.

The function of swapping out specified pages is add.
The process adds special flag for pages to be swapped
out. This function depends on etmem_scan module.

Added swap_flag field to indicate that a tagged swap
out is used for this process instead of the default
swap out.

Signed-off-by: liubo <liubo254@huawei.com>