etmem: fix the problem of libso no permission verification
etmem supports third-party policy injection. During the injection
process, the required so path is specified by the configuration
file.

However, in the so loading process, the lack of effective permission
verification may introduce problems.

Therefore, this patch adds permission verification during the so loading
process, and only allows root permissions to load so.

Signed-off-by: liubo <liubo254@huawei.com>